Update docs and factor out tests.
